[MCH] moved MergeableTH2Ratio class into common MUON folder
