+++ Modules/MUON/CMakeLists.txt
@@ -1,4 +1,5 @@
 # ---- Muon directories ----
 
+add_subdirectory(MUONCommon)
 add_subdirectory(MCH)
 add_subdirectory(MID)
