+++ Modules/MUON/MCH/include/MCH/LinkDef.h
@@ -12,7 +12,6 @@
 #pragma link C++ class o2::quality_control_modules::muonchambers::PhysicsOccupancyCheck + ;
 #pragma link C++ class o2::quality_control_modules::muonchambers::PhysicsPreclustersCheck + ;
 #pragma link C++ class o2::quality_control_modules::muonchambers::TH1MCHReductor + ;
-#pragma link C++ class o2::quality_control_modules::muonchambers::MergeableTH2Ratio + ;
 #pragma link C++ class o2::quality_control_modules::muonchambers::MergeableTH1OccupancyPerDE + ;
 #pragma link C++ class o2::quality_control_modules::muonchambers::MergeableTH1OccupancyPerDECycle + ;
 #pragma link C++ class o2::quality_control_modules::muonchambers::MergeableTH1PseudoEfficiencyPerDE + ;
