+++ Modules/MUON/MUONCommon/include/MUONCommon/LinkDef.h
@@ -1,0 +1,8 @@
+#ifdef __CLING__
+#pragma link off all globals;
+#pragma link off all classes;
+#pragma link off all functions;
+
+#pragma link C++ class o2::quality_control_modules::muonchambers::MergeableTH2Ratio + ;
+
+#endif
